order-service/requirements.txt

requests
fastapi
uvicorn
pydantic-settings

# OpenTelemetry
opentelemetry-distro
# opentelemetry-api
# opentelemetry-sdk
# opentelemetry-instrumentation
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp
# opentelemetry-exporter-otlp-proto-grpc
# opentelemetry-instrumentation-logging

# Logging
structlog
python-json-logger
